Migrated cities list, detailed info and adding new cities to fragments.

+++ Weather/app/src/main/java/com/example/weather/models/CityModel.kt
@@ -6,7 +6,7 @@ import com.example.weather.models.places_model.PlacesModel
 
 data class CityModel(
 
-    val idInDb: Int?,
+    val idInDb: Long?,
 
     val weatherModel: WeatherModel?,
